fix(qsrt): harden publication preflight
